# web/backend/static/openapi.yaml
openapi: 3.1.0
info:
  title: ESA Host Brief Viewer
  version: 0.2.0
  description: 'ESA Incentive Pipeline backend API.


    **Authentication**


    - Host endpoints: require `esa_session` cookie (set via `POST /api/login`). Valid
    24h.

    - Admin endpoints: require `esa_admin_session` cookie (set via `POST /api/admin/login`).
    Valid 1h.


    **Base URL**: `https://esa-hosting.codescales.xyz`


    **AI/Tool discovery**


    - OpenAPI spec (this file): `GET /openapi.yaml`

    - Natural-language context doc: `GET /llms.txt`

    - Interactive UI: `GET /docs`'
  contact:
    name: ESA Incentive Pipeline
    url: https://github.com/BongoEADGC6/esa-incentive-pipeline
paths:
...
  /api/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/runs:
    get:
      tags:
      - runs
      summary: List Runs
      operationId: list_runs_api_runs_get
      parameters:
      - name: stream
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Stream
      - name: window
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Window
      - name: next_hours
        in: query
        required: false
        schema:
          type: number
          default: 0
          title: Next Hours
      - name: marathon
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Marathon
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runs/{slug}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: get_run_api_runs__slug__get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/incentives:
    get:
      tags:
      - incentives
      summary: List Incentives
      operationId: list_incentives_api_incentives_get
      parameters:
      - name: run_slug
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Run Slug
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Status
      - name: category
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Category
      - name: stream
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Stream
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncentiveDTO'
                title: Response List Incentives Api Incentives Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - incentives
      summary: Create Incentive
      operationId: create_incentive_api_incentives_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncentiveCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncentiveDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/login:
    post:
      tags:
      - admin
      summary: Admin Login
      operationId: admin_login_api_admin_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/restore:
    post:
      tags:
      - admin
      summary: Admin Restore
      operationId: admin_restore_api_admin_restore_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/audit:
    get:
      tags:
      - admin
      summary: Admin Audit
      operationId: admin_audit_api_admin_audit_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntry'
                title: Response Admin Audit Api Admin Audit Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/hosts:
    get:
      tags:
      - admin
      summary: Admin List Hosts
      operationId: admin_list_hosts_api_admin_hosts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HostDTO'
                type: array
                title: Response Admin List Hosts Api Admin Hosts Get
    post:
      tags:
      - admin
      summary: Admin Create Host
      operationId: admin_create_host_api_admin_hosts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/hosts/{host_id}:
    delete:
      tags:
      - admin
      summary: Admin Deactivate Host
      operationId: admin_deactivate_host_api_admin_hosts__host_id__delete
      parameters:
      - name: host_id
        in: path
        required: true
        schema:
          type: integer
          title: Host Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/jobs:
    get:
      tags:
      - admin
      summary: List Jobs
      operationId: list_jobs_api_admin_jobs_get
      parameters:
      - name: kind
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Kind
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDTO'
                title: Response List Jobs Api Admin Jobs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/jobs/{id}:
    get:
      tags:
      - admin
      summary: Get Job
      operationId: get_job_api_admin_jobs__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/jobs/{id}/cancel:
    post:
      tags:
      - admin
      summary: Cancel Job
      operationId: cancel_job_api_admin_jobs__id__cancel_post
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/sync/runners:
    post:
      tags:
      - admin
      summary: Sync Runners
      operationId: sync_runners_api_admin_sync_runners_post
      parameters:
      - name: slug
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runners/{slug}:
    get:
      tags:
      - runners
      summary: Get Runner
      operationId: get_runner_api_runners__slug__get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunnerDTOResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: 'Returns runner identity and **current-marathon** run counts only.


        - `run_count` — number of runs in this event''s schedule

        - `esa_count` — distinct streams in this event (not all-time event count)

        - `first_esa` — earliest scheduled run date in the current marathon


        For all-time historical data (17 events, 26 appearances, first debut year,
        country, SRC tenure, ESA run history) call **GET /api/runners/{slug}/profile**.'
...
  /api/runners/{slug}/runs:
    get:
      tags:
      - runners
      summary: Get Runner Runs
      operationId: get_runner_runs_api_runners__slug__runs_get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IncentiveCreateRequest:
      properties:
        run_slug:
          type: string
          title: Run Slug
        incentive_text:
          type: string
          title: Incentive Text
        incentive_category:
          type: string
          title: Incentive Category
          default: ''
        valid_for_game:
          type: string
          title: Valid For Game
          default: ''
        incentive_estimate:
          type: string
          title: Incentive Estimate
          default: ''
        status:
          type: string
          title: Status
          default: ''
      type: object
      required:
      - run_slug
      - incentive_text
      title: IncentiveCreateRequest
      description: Fields required to create a new incentive.
    IncentiveDTO:
      properties:
        scheduled:
          type: string
          format: date-time
          title: Scheduled
        game:
          type: string
          title: Game
        category:
          type: string
          title: Category
        stream:
          type: string
          title: Stream
        runner_display:
          type: string
          title: Runner Display
        runner_twitch:
          type: string
          title: Runner Twitch
        runner_discord:
          type: string
          title: Runner Discord
        runner_slug:
          type: string
          title: Runner Slug
          default: ''
        incentive_text:
          type: string
          title: Incentive Text
        incentive_category:
          type: string
          title: Incentive Category
        valid_for_game:
          type: string
          title: Valid For Game
        incentive_estimate:
          type: string
          title: Incentive Estimate
        needs_approval:
          type: string
          title: Needs Approval
        status:
          type: string
          title: Status
        submission_id:
          type: string
          title: Submission Id
        uuid:
          type: string
          title: Uuid
        run_slug:
          type: string
          title: Run Slug
          default: ''
        participants:
          items:
            $ref: '#/components/schemas/ParticipantDTO'
          type: array
          title: Participants
          default: []
      type: object
      required:
      - scheduled
      - game
      - category
      - stream
      - runner_display
      - runner_twitch
      - runner_discord
      - incentive_text
      - incentive_category
      - valid_for_game
      - incentive_estimate
      - needs_approval
      - status
      - submission_id
      - uuid
      title: IncentiveDTO
...
    LoginRequest:
      properties:
        password:
          type: string
          title: Password
      type: object
      required:
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        ok:
          type: boolean
          title: Ok
      type: object
      required:
      - ok
      title: LoginResponse
...
    RestoreRequest:
      properties:
        snapshot_id:
          type: string
          title: Snapshot Id
      type: object
      required:
      - snapshot_id
      title: RestoreRequest
...
  securitySchemes:
    hostSession:
      type: apiKey
      in: cookie
      name: esa_session
      description: Set via POST /api/login. Valid 24h.
    adminSession:
      type: apiKey
      in: cookie
      name: esa_admin_session
      description: Set via POST /api/admin/login. Valid 1h.
servers:
- url: https://esa-hosting.codescales.xyz
  description: Production
...